arreglo de Configuracion de crupier y jugador
